Patch File objects sent as form data to work around a bug in Chrome (#36)
* Patch File objects sent as form data to work around a bug in Chrome

Before sending request data over `window.postMessage` to the proxy iframe, we find all `File`
instances in the form data and create a new `File` instance based on the existing `File`. That
forces the `File` object's storage to be a `Blob` instead of being backed by a file on disk.
That works around a bug in Chrome where `File` instances with `has_backing_file` flag cannot
be sent over a process boundary when site isolation is on.

This PR also removes an old workaround for Firefox where the browser couldn't send a `File` over
`window.postMessage` at all. That bug has been fixed for several years now.

* Feature detect File constructor on IE and Edge

* Update package-lock.json