Fixes touch device bug with inability to select range in DateRange component
Previously touch devices were unable to select a range of dates. Instead when attempting to select a range the same date would be selected to become both the from and the to dates. This was because the underlying DatePicker component whereby the handler method was firing multiple times  due to it being assigned to both touch and click events.

The result of this was that the handler would be called once for touch and then again for click resulting in any method passed into DatePicker as `this.props.onSelectDay` being fired mutliple times. If - like in DataRange - that method was used to set the selected dates on the DatePicker then it would cause the behaviour described above.

Removing these additional touch handlers fixes the issue. It was originally introduced in this commit (https://github.com/Automattic/wp-calypso/pull/27287/commits/406a5f6226761b5ef3ff17f35f5d42956880ea6e) but no explaination was given as to _why_ it was needed.