Fix date range class to be more robustly applied depending on modifiers
Previously the date range class `date-picker—has-range` was applied based on prescence of from or to props on the selectedDays prop. However we cannot rely on this format as ranges can be passed in a variety of formats.

The most reliable method is to rely on a start/end properties within the `modifiers` prop. Alter Date Picker to conform to this.